+++ tomcat/src/main/java/org/apache/catalina/ServletContainer.java
@@ -1,7 +1,6 @@
-package org.apache.catalina.http11;
+package org.apache.catalina;
 
 import java.util.Map;
-import org.apache.catalina.http.HeaderName;
 import org.apache.catalina.controller.Controller;
 import org.apache.catalina.controller.HomeController;
 import org.apache.catalina.controller.LoginController;
@@ -9,11 +8,11 @@ import org.apache.catalina.controller.RegisterController;
 import org.apache.catalina.request.HttpRequest;
 import org.apache.catalina.response.HttpResponse;
 
-public class Dispatcher {
+public class ServletContainer {
 
     private final Map<String, Controller> controllers;
 
-    public Dispatcher() {
+    public ServletContainer() {
         this.controllers = Map.of(
                 "/login", new LoginController(),
                 "/register", new RegisterController(),
@@ -22,11 +21,6 @@ public class Dispatcher {
     }
 
     public void run(HttpRequest httpRequest, HttpResponse httpResponse) {
-        httpResponse.addHeader(HeaderName.CONTENT_TYPE, httpRequest.getContentType());
-        if (httpRequest.hasCookie()) {
-            httpResponse.addHeader(HeaderName.SET_COOKIE, httpRequest.getHttpCookie());
-        }
-
         Controller controller = controllers.get(httpRequest.getPath());
         controller.service(httpRequest, httpResponse);
     }
